Refactor Tasks.ts by splitting into separate files
* Remove TaskDecorator was they are not working without importing the task files
