+++ src/tasks/utils.ts
@@ -1,0 +1,57 @@
+import * as d3 from 'd3';
+
+// SOURCE: https://stackoverflow.com/a/51592360/2549748
+/**
+ * Deep copy function for TypeScript.
+ * @param T Generic type of target/copied value.
+ * @param target Target value to be copied.
+ * @see Source project, ts-deepcopy https://github.com/ykdr2017/ts-deepcopy
+ * @see Code pen https://codepen.io/erikvullings/pen/ejyBYg
+ */
+export function deepCopy<T>(target: T): T {
+  if (target === null) {
+    return target;
+  }
+  if (target instanceof Date) {
+    return new Date(target.getTime()) as any;
+  }
+  if (target instanceof Array) {
+    const cp = [] as any[];
+    (target as any[]).forEach((v) => { cp.push(v); });
+    return cp.map((n: any) => deepCopy<any>(n)) as any;
+  }
+  if (typeof target === 'object' && target !== {}) {
+    const cp = { ...(target as { [key: string]: any }) } as { [key: string]: any };
+    Object.keys(cp).forEach((k) => {
+      cp[k] = deepCopy<any>(cp[k]);
+    });
+    return cp as T;
+  }
+  return target;
+}
+
+export function score2color(score: number): { background: string, foreground: string } {
+  let background = '#ffffff'; // white
+  let foreground = '#333333'; // kinda black
+
+
+  if (score <= 0.05) {
+    // console.log('bg color cahnge')
+    const calcColor = d3.scale.linear().domain([0, 0.05]).range(<any[]>['#000000', '#FFFFFF']);
+
+    background = calcColor(score).toString();
+    foreground = textColor4Background(background);
+  }
+
+  return { background, foreground };
+}
+
+
+export function textColor4Background(backgroundColor: string) {
+  let color = '#333333';
+  if ('transparent' !== backgroundColor && d3.hsl(backgroundColor).l < 0.5) { // transparent has lightness of zero
+    color = 'white';
+  }
+
+  return color;
+}
